+++ CHANGELOG.md
@@ -22,6 +22,11 @@
   [Ornithologist Coder](https://github.com/ornithocoder)
   [#1874](https://github.com/realm/SwiftLint/issues/1874)
 
+* Add `last_where` opt-in rule that warns against using
+  `.filter { /* ... */ }.last` in collections, as
+  `.last(where: { /* ... */ })` is more efficient.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+
 #### Bug Fixes
 
 * None.